+++ RELEASE-NOTES.md
@@ -19,7 +19,7 @@
 ## Deprecations
 
 ## New additions
-* `snow streamlit deploy`: add `--follow-symlinks` flag to allow following symlinks that point outside the project root ([SNOW-3417049](https://snowflakecomputing.atlassian.net/browse/SNOW-3417049))
+* `snow streamlit deploy`: add `--follow-symlinks` flag to allow following symlinks that point outside the project root.
 
 ## Fixes and improvements